move lease related code into a separate file lease.go
